+++ Authentication/Token/NullToken.php
@@ -43,14 +43,6 @@ class NullToken implements TokenInterface
         return '';
     }
 
-    /**
-     * @deprecated since Symfony 7.3
-     */
-    #[\Deprecated(since: 'symfony/security-core 7.3')]
-    public function eraseCredentials(): void
-    {
-    }
-
     public function getAttributes(): array
     {
         return [];